% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProKlust.R
\name{plotc}
\alias{plotc}
\title{Plot the computed clusters igraph.}
\usage{
plotc(igraph, plotFontSize = 10, plotLinkDistance = 60, nodeAttraction = -10)
}
\arguments{
\item{igraph}{Obligatory igraph object.}

\item{plotFontSize}{Optional numeric font size in pixels for the node text labels.}

\item{plotLinkDistance}{Optional numeric or character string. Either numberic fixed distance between the links in pixels (actually arbitrary relative to the diagram's size). Or a JavaScript function, possibly to weight by Value. For example: linkDistance = JS("function(d){return d.value * 10}").}

\item{nodeAttraction}{Optional numeric value indicating either the strength of the node repulsion (negative value) or attraction (positive value)}
}
\value{
The actual plot of the network (an igraph which was transformed into graphical plot by networkD3).
}
\description{
networkD3 library is used to plot the graph and a HTML copy file named network.html is created.
}
\examples{
files <- c("ANIb_percentage_identity.tab", "ANIb_alignment_coverage.tab")
thresholds <- c(0.95, 0.70)
results <- prokluster(filenames = files, cutoffs = thresholds, nodesDictionary = "dictionary.tab")
plottedD3 <- plotc(results$graph)

}
